requirements.txt
-i https://pypi.org/simple
cachecontrol==0.12.11 ; python_version >= '3.6'
cachetools==5.2.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1 ; python_version >= '3.6'
click==8.1.3 ; python_version >= '3.7'
cryptography==39.0.0
dateparser==1.1.5
envelopes==0.4
firebase-admin==6.0.1
flask==2.2.2
google-api-core[grpc]==2.11.0 ; platform_python_implementation != 'PyPy'
google-api-python-client==2.70.0 ; python_version >= '3.7'
google-auth==2.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.2.0 ; python_version >= '3.7'
google-cloud-audit-log==0.2.4 ; python_version >= '3.7'
google-cloud-bigquery==3.4.1
google-cloud-core==2.3.2 ; python_version >= '3.7'
google-cloud-firestore==2.7.3 ; platform_python_implementation != 'PyPy'
google-cloud-logging==3.4.0
google-cloud-pubsub==2.13.11
google-cloud-secret-manager==2.13.0
google-cloud-storage==2.7.0 ; python_version >= '3.7'
google-cloud-tasks==2.11.0
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.4.0 ; python_version >= '3.7'
googleapis-common-protos==1.57.0 ; python_version >= '3.7'
grpc-google-iam-v1==0.12.4 ; python_version >= '3.6'
grpcio==1.51.1 ; python_version >= '3.7'
grpcio-status==1.51.1 ; python_version >= '3.6'
httplib2==0.21.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2
markupsafe==2.1.1 ; python_version >= '3.7'
msgpack==1.0.4
numpy==1.24.1
packaging==21.3 ; python_version >= '3.6'
proto-plus==1.22.1 ; python_version >= '3.6'
protobuf==4.21.12 ; python_version >= '3.7'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyjwt[crypto]==2.6.0 ; python_version >= '3.7'
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
regex==2022.10.31 ; python_version >= '3.6'
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
rsa==4.9 ; python_version >= '3.6'
sentry-sdk==1.12.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tinify==1.6.0
tzdata==2022.7 ; python_version >= '3.6'
tzlocal==4.2
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.13 ; python_version >= '3.6'
werkzeug==2.2.2 ; python_version >= '3.7'
zipp==3.11.0 ; python_version >= '3.7'
